main: generate towns via buildings_v2 generate_town pipeline
Extract the full settlement pipeline (parcels -> wall -> flatten -> industrial -> tiered A* roads -> subdivision -> houses -> verge + street lights) out of the hierarchical_roads test into a reusable generate_town(editor, seed, culture) in generator/settlement.rs. main's run_generation_once now calls it with Culture::Medieval, retiring the old buildings v1 place_buildings path. The test is now a thin wrapper so it and main can't drift.
