+++ src/generator/districts/test.rs
@@ -1,4 +1,4 @@
-#[cfg(test)]
+﻿#[cfg(test)]
 mod tests {
     use std::collections::{HashMap, HashSet};
     use crate::{data::Loadable, editor::World, generator::districts::{WallType, build_wall, HasParcelData, parcel::{self, generate_parcels}, parcel_painter::{replace_ground, replace_ground_smooth}}, geometry::{Point2D, Point3D}, http_mod::{GDMCHTTPProvider, HeightMapType}, minecraft::Block, noise::{RNG, Seed}, util::init_logger};
@@ -227,7 +227,7 @@ mod tests {
         let pole: Block = "oak_fence".into();
         for (id, parcel_type, centre, size) in sign_info {
             if centre.x < 0 || centre.y < 0 || centre.x >= build_area.size.x || centre.y >= build_area.size.z {
-                log::info!("Super-parcel {} final type={:?} size={} cells, centre={:?} out of bounds — no sign", id, parcel_type, size, centre);
+                log::info!("Super-parcel {} final type={:?} size={} cells, centre={:?} out of bounds â€” no sign", id, parcel_type, size, centre);
                 continue;
             }
 
@@ -235,7 +235,7 @@ mod tests {
             let surface_y = height_map[centre.x as usize][centre.y as usize];
             let (world_x, world_z) = (centre.x + build_area.origin.x, centre.y + build_area.origin.z);
             log::info!(
-                "Super-parcel {} final type={:?} size={} cells — sign at world ({}, {}, {})  /tp @s {} {} {}",
+                "Super-parcel {} final type={:?} size={} cells â€” sign at world ({}, {}, {})  /tp @s {} {} {}",
                 id, parcel_type, size, world_x, surface_y + 4, world_z, world_x, surface_y + 5, world_z
             );
 
@@ -247,7 +247,7 @@ mod tests {
             editor.place_block(&sign_block(&id.to_string()), Point3D::new(centre.x, h + 4, centre.y)).await;
         }
 
-        // Verify the size band: every Urban/Rural (interior) parcel should be within ±50% of the
+        // Verify the size band: every Urban/Rural (interior) parcel should be within Â±50% of the
         // interior average block count. Off-limits parcels are exempt. See
         // docs/plans/parcel_size_balancing.md.
         let interior_sizes: Vec<(usize, parcel::ParcelType, usize)> = editor.world()
@@ -417,7 +417,7 @@ mod tests {
             "oak_wood_roof".into(),
             "red_wood_roof".into(),
         ];
-        // Medieval-feeling woods and stones — skipping tropical (jungle, acacia)
+        // Medieval-feeling woods and stones â€” skipping tropical (jungle, acacia)
         // and nether (blackstone) variants. Each is a partial palette that only
         // overrides its respective roles, so stacking them gives ~36 combos.
         let wood_palette_ids: Vec<PaletteId> = vec![
@@ -462,7 +462,7 @@ mod tests {
 
         // All road cells: perimeter + alleys. Claim as PathPlanned so the
         // frontage walker treats them as roads, but foundation terrain
-        // blending will still raise the heightmap on them — meaning the
+        // blending will still raise the heightmap on them â€” meaning the
         // post-house pave step picks up the foundation-influenced heights.
         let road_cells: HashSet<Point2D> = perimeter_roads.iter().chain(all_alleys.iter()).copied().collect();
         for p in &road_cells {
@@ -541,7 +541,7 @@ mod tests {
                 None => continue,
             };
 
-            // Frontage pass — one house per slot along each frontage chain.
+            // Frontage pass â€” one house per slot along each frontage chain.
             let frontages = {
                 let detected = detect_frontages(sub_block, &editor);
                 if detected.is_empty() {
@@ -598,7 +598,7 @@ mod tests {
                 }
             }
 
-            // Interior pass disabled — frontage only for now.
+            // Interior pass disabled â€” frontage only for now.
             // let max_interior = 10usize;
             // let mut placed = 0usize;
             // while placed < max_interior {
@@ -640,7 +640,7 @@ mod tests {
 
         editor.flush_buffer().await;
         println!(
-            "Done — {} total buildings across {} sub-blocks",
+            "Done â€” {} total buildings across {} sub-blocks",
             total_buildings, all_sub_blocks.len(),
         );
     }
@@ -773,9 +773,9 @@ mod tests {
         let mut producing_ids: Vec<_> = result.parcel_assignments.keys().cloned().collect();
         producing_ids.sort_by_key(|id| id.0);
 
-        println!("\n╔══ Parcel Resource Production Report ════════════╗");
+        println!("\nâ•”â•â• Parcel Resource Production Report â•â•â•â•â•â•â•â•â•â•â•â•â•—");
 
-        println!("║ Producing Super-Parcels ({} rural of {} total):", producing_ids.len(), editor.world().district_analysis_data.len());
+        println!("â•‘ Producing Super-Parcels ({} rural of {} total):", producing_ids.len(), editor.world().district_analysis_data.len());
         for id in &producing_ids {
             let analysis = &editor.world().district_analysis_data[id];
             let biome_names = {
@@ -786,50 +786,50 @@ mod tests {
                 names.join("+")
             };
             let a = &result.parcel_assignments[id];
-            println!("║   Super-Parcel {:>3} ({:<25}) → {} x2 [{}]",
+            println!("â•‘   Super-Parcel {:>3} ({:<25}) â†’ {} x2 [{}]",
                 id.0, biome_names, a.primary_resource, a.building);
         }
 
-        println!("║");
-        println!("║ Resource Supply:");
+        println!("â•‘");
+        println!("â•‘ Resource Supply:");
         let mut supply_sorted: Vec<(&String, &u32)> = result.supply.iter().collect();
         supply_sorted.sort_by_key(|(r, _)| r.as_str());
         for (resource, qty) in supply_sorted {
-            println!("║   {:<20} x{}", resource, qty);
+            println!("â•‘   {:<20} x{}", resource, qty);
         }
 
-        println!("║");
-        println!("║ Goods Produced:");
+        println!("â•‘");
+        println!("â•‘ Goods Produced:");
         if result.finished_goods.is_empty() && result.leftover_goods.is_empty() {
-            println!("║   (none)");
+            println!("â•‘   (none)");
         }
         for (good, qty) in &result.finished_goods {
-            println!("║   {:<20} x{}", good, qty);
+            println!("â•‘   {:<20} x{}", good, qty);
         }
         for (good, qty) in &result.leftover_goods {
-            println!("║   {:<20} x{}  (unused)", good, qty);
+            println!("â•‘   {:<20} x{}  (unused)", good, qty);
         }
 
-        println!("║");
-        println!("║ Gathering Buildings:");
+        println!("â•‘");
+        println!("â•‘ Gathering Buildings:");
         let mut gb_sorted: Vec<(&String, &u32)> = result.gather_buildings.iter().collect();
         gb_sorted.sort_by_key(|(b, _)| b.as_str());
         for (building, count) in gb_sorted {
-            println!("║   {:<20} x{}", building, count);
+            println!("â•‘   {:<20} x{}", building, count);
         }
 
-        println!("║");
-        println!("║ Processing Buildings Required:");
+        println!("â•‘");
+        println!("â•‘ Processing Buildings Required:");
         if result.processing_buildings.is_empty() {
-            println!("║   (none)");
+            println!("â•‘   (none)");
         }
         let mut pb_sorted: Vec<(&String, &u32)> = result.processing_buildings.iter().collect();
         pb_sorted.sort_by(|(a, ac), (b, bc)| bc.cmp(ac).then(a.cmp(b)));
         for (building, count) in pb_sorted {
-            println!("║   {:<20} x{}", building, count);
+            println!("â•‘   {:<20} x{}", building, count);
         }
 
-        println!("╚═══════════════════════════════════════════════════╝\n");
+        println!("â•šâ•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•â•\n");
     }
 
     #[tokio::test]
@@ -1129,580 +1129,14 @@ mod tests {
     /// parcels -> wall+gates -> flatten -> arterials(MST)+collectors(gates) -> build_path.
     #[tokio::test]
     async fn hierarchical_roads() {
-        use crate::generator::data::LoadedData;
-        use crate::generator::paths::{build_paths_merged, build_road_network, find_blocks, Path, PathPriority};
-        use crate::generator::terrain::{flatten_urban_area, force_height, log_trees};
-
         init_logger();
-
-        let seed = Seed(12345);
-        let mut rng = RNG::new(seed);
-        let mut rng2 = RNG::new(seed);
-
         let provider = GDMCHTTPProvider::new();
         let world = World::new(&provider).await.expect("Failed to create world");
         let mut editor = world.get_editor();
-
-        generate_parcels(seed, &mut editor).await;
-
-        // EVAL AID (test-only): the live server hands out a different build area
-        // each run, and the urban classifier frequently collapses to a single
-        // parcel — too degenerate to evaluate the road network. Force a
-        // contiguous ~4-parcel urban core: keep the prime, then promote the
-        // nearest non-off-limits super-parcels to Urban.
-        {
-            use crate::generator::districts::ParcelType;
-            const TARGET_URBAN: usize = 4;
-            let mut info: Vec<(crate::generator::districts::DistrictID, Point2D, bool)> =
-                editor.world().districts.iter()
-                    .filter(|(id, sd)| {
-                        if sd.data.parcel_type == ParcelType::OffLimits {
-                            return false;
-                        }
-                        // Never force a water-heavy parcel urban — it would build the
-                        // town on a lake. (Matches URBAN_WATER_LIMIT in classification.)
-                        let water = editor.world().district_analysis_data
-                            .get(id)
-                            .map_or(0.0, |a| a.water_percentage());
-                        water <= 0.33
-                    })
-                    .map(|(id, sd)| {
-                        let pts = &sd.data.points_2d;
-                        let c = pts.iter().fold(Point2D::ZERO, |a, p| a + *p) / pts.len().max(1) as i32;
-                        (*id, c, sd.data.parcel_type == ParcelType::Urban)
-                    })
-                    .collect();
-            let anchor = info.iter().find(|(_, _, u)| *u).map(|(_, c, _)| *c)
-                .or_else(|| info.first().map(|(_, c, _)| *c));
-            if let Some(anchor) = anchor {
-                info.sort_by_key(|(_, c, _)| c.distance_squared(&anchor));
-                for (id, _, _) in info.iter().take(TARGET_URBAN) {
-                    editor.world_mut().districts.get_mut(id).unwrap().data.parcel_type = ParcelType::Urban;
-                }
-            }
-        }
-
-        // Wall + gates — gates populate world.gate_locations, used by the network.
-        let materials = Material::load().expect("Failed to load materials");
-        let wall_material = MaterialId::new("stone_bricks".to_string());
-        let mut placer: Placer = Placer::new(&materials, &mut rng);
-        let structures = Structure::load().expect("Failed to load structures");
-        build_wall(
-            &editor.world().get_urban_points(), &mut editor, &mut rng2,
-            &mut placer, &wall_material, &structures, WallType::Standard,
+        crate::generator::settlement::generate_town(
+            &mut editor,
+            Seed(12345),
+            crate::generator::buildings_v2::Culture::Medieval,
         ).await;
-        drop(placer);
-
-        // DEBUG: how many urban super-parcels and gates did we actually get?
-        {
-            let n_urban = editor.world().districts.values()
-                .filter(|sd| sd.data.parcel_type == crate::generator::districts::ParcelType::Urban)
-                .count();
-            let n_total = editor.world().districts.len();
-            println!("URBAN super-parcels: {}/{} total | gates: {}", n_urban, n_total, editor.world().gate_locations.len());
-        }
-
-        // Phase 1 — feathered urban flatten.
-        let urban = editor.world().get_urban_points();
-        // Log (clear) the urban area of trees so roads, buildings, and houses
-        // aren't dropped into standing forest.
-        log_trees(&editor, urban.clone()).await;
-        println!("Logged {} urban cells of trees", urban.len());
-        flatten_urban_area(&mut editor, &urban, 16, 12, true).await;
-
-        let data = LoadedData::load().expect("Failed to load data");
-
-        // ---- Industrial buildings FIRST ----
-        // Place a handful of big processing buildings on the flattened ground (no
-        // roads yet → sited by flatness). They become the destinations the arterial
-        // network connects, plus a `blocked` barrier so nothing — roads, the
-        // subdivision, alleys, or houses — ever runs through them. (Fixed set here;
-        // the resource chain's `resolve_for_parcels` can supply the real mix later.)
-        use crate::generator::BuildClaim;
-        use crate::generator::placement::place_urban_buildings;
-
-        let mut ind_counts: HashMap<String, u32> = HashMap::new();
-        for b in ["smithy", "mill", "bakery", "carpenter", "tannery", "weaver"] {
-            ind_counts.insert(b.to_string(), 1);
-        }
-        let urban_sds: Vec<_> = editor.world().districts.values()
-            .filter(|sd| sd.data.parcel_type == crate::generator::districts::ParcelType::Urban)
-            .cloned()
-            .collect();
-        let urban_sd_refs: Vec<_> = urban_sds.iter().collect();
-        let n_before = editor.world().structures.len();
-        // Re-skin the industrial NBTs into the settlement's culture palette
-        // (their baked `resource_base` blocks → medieval spruce/stone).
-        let ind_palette = data.palettes
-            .get(&crate::generator::buildings_v2::Culture::Medieval.palette_id())
-            .expect("industry palette not found").clone();
-        if let Err(e) = place_urban_buildings(&urban_sd_refs, &ind_counts, &mut rng, &mut editor, &data, Some(&ind_palette)).await {
-            log::warn!("industrial placement failed: {}", e);
-        }
-        println!(
-            "Placed {} / {} industrial buildings",
-            editor.world().structures.len() - n_before, ind_counts.values().sum::<u32>(),
-        );
-
-        // Footprints → a `blocked` barrier (footprint + margin) and one node per
-        // building for the network to connect.
-        const IND_MARGIN: i32 = 2;
-        let mut ind_footprints: HashMap<u32, Vec<Point2D>> = HashMap::new();
-        for &p in &urban {
-            if let Some(BuildClaim::Structure(id)) = editor.world().get_claim(p) {
-                ind_footprints.entry(id.id).or_default().push(p);
-            }
-        }
-        let building_cells: HashSet<Point2D> = ind_footprints.values().flatten().copied().collect();
-        let blocked: HashSet<Point2D> = building_cells.iter()
-            .flat_map(|p| {
-                (-IND_MARGIN..=IND_MARGIN).flat_map(move |dx| {
-                    (-IND_MARGIN..=IND_MARGIN).map(move |dz| Point2D::new(p.x + dx, p.y + dz))
-                })
-            })
-            .collect();
-        let ind_nodes: Vec<Point3D> = ind_footprints.values()
-            .map(|cells| {
-                let c = cells.iter().fold(Point2D::ZERO, |a, p| a + *p) / cells.len().max(1) as i32;
-                editor.world().add_height(c)
-            })
-            .collect();
-
-        // Phase 2 — tiered A* road network, connecting the industrial buildings
-        // (anchor nodes) and routed around them (the `blocked` barrier).
-        let arterial_material = MaterialId::new("stone_bricks".to_string());
-        let collector_material = MaterialId::new("cobblestone".to_string());
-        let paths = build_road_network(
-            &editor, arterial_material, collector_material, true, &ind_nodes, &blocked, 1,
-        ).await.paths;
-        println!("Routed {} road segments", paths.len());
-
-        // DEBUG: Phase A merge check — how many of each path's cells coincide
-        // with cells already laid by earlier paths? High overlap = routes are
-        // merging onto the network instead of crossing it blindly.
-        {
-            let mut seen: HashSet<Point2D> = HashSet::new();
-            for path in &paths {
-                let cells: HashSet<Point2D> = path.points().iter().map(|p| p.drop_y()).collect();
-                let shared = cells.iter().filter(|c| seen.contains(c)).count();
-                println!("  MERGE prio={:?} pts={} shared_with_network={}", path.priority(), cells.len(), shared);
-                seen.extend(cells);
-            }
-        }
-
-        // DEBUG: does the routed path y match the post-flatten heightmap?
-        if let Some(path) = paths.first() {
-            println!("--- path[0] sample: road_y vs ground_h vs ocean_h ---");
-            for p in path.points().iter().take(25) {
-                let xz = p.drop_y();
-                println!(
-                    "  ({:>4},{:>4})  road_y={:>3}  ground_h={:>3}  ocean_h={:>3}",
-                    xz.x, xz.y, p.y,
-                    editor.world().get_height_at(xz),
-                    editor.world().get_ocean_floor_height_at(xz),
-                );
-            }
-        }
-
-        // A path's *paved* cells — exactly what `build_paths_merged` lays:
-        // centreline ∪ (width-1) ring. Used for block barriers and frontage
-        // bands so blocks abut the real road edge (no gap ring).
-        let paved = |path: &Path| -> HashSet<Point2D> {
-            let centre: HashSet<Point2D> = path.points().iter().map(|p| p.drop_y()).collect();
-            let mut cells = crate::geometry::get_surrounding_set(&centre, path.width().saturating_sub(1));
-            cells.extend(centre);
-            cells
-        };
-
-        // Blocks = urban minus the paved main roads and the wall.
-        let wall: HashSet<Point2D> = urban.iter()
-            .filter(|&&c| crate::geometry::CARDINALS_2D.iter().any(|&d| !urban.contains(&(c + d))))
-            .copied()
-            .collect();
-        let mut barriers: HashSet<Point2D> = HashSet::new();
-        for path in &paths {
-            barriers.extend(paved(path));
-        }
-        barriers.extend(&wall);
-        // Industrial buildings (footprint + margin) are barriers too, so blocks —
-        // and the subdivision, alleys, and houses inside them — form *around* the
-        // buildings, never through them.
-        barriers.extend(&blocked);
-
-        // Don't let blocks (and the lots/alleys/houses inside them) span steep
-        // terrain. A per-cell cliff test misses a *sustained* slope — a long
-        // staircase of 1-block risers passes cell-by-cell yet climbs far. So bar
-        // any cell whose local WIN-radius neighbourhood spans more than
-        // MAX_LOCAL_RELIEF blocks of height; the flood fill then breaks blocks
-        // along slope lines, keeping lots and their lanes on a flat shelf.
-        const WIN: i32 = 1; // 3×3 window
-        const MAX_LOCAL_RELIEF: i32 = 2;
-        let steep: HashSet<Point2D> = urban.iter()
-            .filter(|&&c| {
-                let (mut lo, mut hi) = (i32::MAX, i32::MIN);
-                for dx in -WIN..=WIN {
-                    for dz in -WIN..=WIN {
-                        let n = Point2D::new(c.x + dx, c.y + dz);
-                        if !urban.contains(&n) { continue; }
-                        let h = editor.world().get_ocean_floor_height_at(n);
-                        lo = lo.min(h);
-                        hi = hi.max(h);
-                    }
-                }
-                hi - lo > MAX_LOCAL_RELIEF
-            })
-            .copied()
-            .collect();
-        println!("Marked {} steep cells as barriers", steep.len());
-        barriers.extend(&steep);
-
-        let blocks = find_blocks(&urban, &barriers, 12);
-        println!("Found {} blocks", blocks.len());
-
-        // All main-road (arterial + collector) paved cells, used to peel a
-        // frontage ribbon off each block before subdividing its interior.
-        let main_road_cells: HashSet<Point2D> = {
-            let mut s = HashSet::new();
-            for path in &paths {
-                s.extend(paved(path));
-            }
-            s
-        };
-
-        // Per block: first reserve a frontage ribbon — a band one house deep
-        // against each main road — so the long arterial/collector-facing edge
-        // stays a single continuous lot instead of being chopped into stubs
-        // by subdivision. Then subdivide only the interior with tier-3 alleys.
-        // BSP cuts span the interior edge-to-edge, so an alley reaches its edge —
-        // adjacent (barriers = paved) to either a main road or the ribbon.
-        // Deep enough to absorb both the deepest House (depth_range 7..=10) AND
-        // the staircase rise of a diagonal frontage (an axis-aligned rect anchored
-        // at the slice's interior extreme reaches `rise + depth` into the band).
-        const RIBBON_DEPTH: i32 = 14;
-        let mut sub_blocks: Vec<HashSet<Point2D>> = Vec::new();
-        let mut alley_band: HashSet<Point2D> = HashSet::new();
-        let mut ribbon_lot_count = 0usize;
-        let mut ribbon_cells: HashSet<Point2D> = HashSet::new(); // DEBUG: all reserved ribbon cells
-        for block in &blocks {
-            let (mut ribbon_lots, interior) =
-                crate::generator::districts::subdivide::reserve_road_ribbon(block, &main_road_cells, RIBBON_DEPTH);
-            let (subs, alleys) = crate::generator::districts::subdivide::subdivide_block(&interior, &mut rng, 24);
-
-            // Connect the interior alleys to the main roads by carving through the
-            // ribbon, then convert those cells from frontage ribbon to alley.
-            let ribbon_union: HashSet<Point2D> = ribbon_lots.iter().flatten().copied().collect();
-            let connectors = crate::generator::districts::subdivide::carve_ribbon_connectors(
-                &ribbon_union, &alleys, &main_road_cells,
-            );
-            if !connectors.is_empty() {
-                for rp in &mut ribbon_lots { rp.retain(|c| !connectors.contains(c)); }
-                ribbon_lots.retain(|rp| !rp.is_empty());
-            }
-
-            ribbon_lot_count += ribbon_lots.len();
-            for rp in &ribbon_lots { ribbon_cells.extend(rp); }
-            sub_blocks.extend(ribbon_lots);
-            alley_band.extend(&alleys);
-            alley_band.extend(&connectors);
-            sub_blocks.extend(subs);
-        }
-        println!(
-            "Subdivided into {} lots ({} road-frontage ribbons), {} subdivider-road cells",
-            sub_blocks.len(), ribbon_lot_count, alley_band.len(),
-        );
-
-        // Assemble every road into one path list (mains + a synthesised width-1
-        // alley path), but DON'T build them yet — we build after the houses so
-        // house-foundation earth can't bury the road. Houses are placed first and
-        // sit their floor at the level of the road they front (see `road_h`).
-        let alley_pts: Vec<Point3D> = alley_band.iter().map(|c| editor.world().add_height(*c)).collect();
-        let alley_path = Path::new(alley_pts, 1, MaterialId::new("cobblestone".to_string()), PathPriority::Low);
-        let mut all_paths = paths.clone();
-        all_paths.push(alley_path);
-
-        // Road-height lookup over the paved band of every road (centreline +
-        // width ring, min y on overlap), so a house can pin its floor to the
-        // road it fronts. Built from `all_paths` so alley-facing houses get the
-        // alley level too.
-        let mut road_h: HashMap<Point2D, i32> = HashMap::new();
-        for path in &all_paths {
-            let w = path.width() as i32;
-            for pt in path.points() {
-                let base = pt.drop_y();
-                for dx in -w..=w {
-                    for dz in -w..=w {
-                        let c = Point2D::new(base.x + dx, base.y + dz);
-                        road_h.entry(c).and_modify(|y| *y = (*y).min(pt.y)).or_insert(pt.y);
-                    }
-                }
-            }
-        }
-
-        // Frontage bands per tier (paved cells, matching the roads we'll build).
-        let band = |prio: PathPriority| -> HashSet<Point2D> {
-            let mut s = HashSet::new();
-            for path in paths.iter().filter(|p| p.priority() == prio) {
-                s.extend(paved(path));
-            }
-            s
-        };
-        let arterial_band = band(PathPriority::High);
-        let collector_band = band(PathPriority::Medium);
-
-        // Build the roads FIRST, then the houses. force_height grades the corridor
-        // to the routed road heights, then build_paths_merged lays + melds the
-        // surface. We then claim every paved cell as `Path` so the following
-        // house foundations' terrain blending skips them (blend_terrain ignores
-        // Path claims) — the road can't be buried by foundation earth. The graded
-        // corridor is exactly `road_h` (same band, same min-on-overlap height).
-        let corridor_pts: HashSet<Point3D> = road_h
-            .iter()
-            .map(|(c, &y)| Point3D::new(c.x, y, c.y))
-            .collect();
-        force_height(&mut editor, &corridor_pts, false).await;
-        // `build_paths_merged` returns the exact cells where it laid a half-step
-        // slab; we raise a house a block over a fronting slab off this set rather
-        // than reading the placed road back (the editor cache is keyed by local
-        // coords while get_block subtracts the build-area origin, so a read here
-        // returns world terrain, not the road).
-        let road_slabs: HashSet<Point3D> = build_paths_merged(&editor, &data, &all_paths, &mut rng).await;
-        let slab_y_by_cell: HashMap<Point2D, i32> =
-            road_slabs.iter().map(|p| (p.drop_y(), p.y)).collect();
-
-        // Claim every paved road cell so house-foundation terraforming can't
-        // touch it (blend_terrain skips `BuildClaim::Path`).
-        for path in &all_paths {
-            for c in paved(path) {
-                editor.world_mut().claim(c, crate::generator::BuildClaim::Path(crate::generator::paths::PathType::Pavement));
-            }
-        }
-
-        // ---- Phase 4: hierarchical house placement ----
-        // Per lot, walk frontage densest-tier first: arterial → collector →
-        // subdivider. The lot's single Plot is shared across tiers, so houses
-        // placed against the arterial claim the prime frontage and later tiers
-        // can't overlap them. Size gradient: houses on roads, cottages on lanes.
-        use crate::generator::buildings_v2::{BuildCtx, BuildingContext, Culture, build_house};
-        use crate::generator::buildings_v2::roof::RoofStyle;
-        use crate::generator::buildings_v2::roof::gable::GablePitch;
-        use crate::generator::buildings_v2::footprint::{Footprint, SizeClass};
-        use crate::generator::city_houses::{
-            frontage_from_roads, plot_from_block, rect_from_frontage,
-            synthetic_plot_bounds, SIDE_BUFFER_CELLS,
-        };
-        use crate::generator::materials::{Palette, PaletteId};
-        use crate::geometry::Point2D as P2;
-
-        // Culture for this settlement. Medieval → spruce/stone palette, gable
-        // roofs, glass windows, and timber-frame jetties (square_bias 0, so no
-        // domes; per-house wood/stone/roof variety via roll_palette below).
-        let culture = Culture::Medieval;
-        let base_palette: Palette = data.palettes.get(&culture.palette_id())
-            .expect("base palette not found").clone();
-        let wood_ids: Vec<PaletteId> = vec!["oak".into(), "spruce".into(), "dark_oak".into()];
-        let stone_ids: Vec<PaletteId> = vec!["stone_bricks".into(), "cobblestone".into(), "deepslate".into()];
-        let roof_ids: Vec<PaletteId> = vec![
-            "acacia_wood_roof".into(), "brick_roof".into(), "oak_wood_roof".into(), "red_wood_roof".into(),
-        ];
-        let roof_styles = culture.roof_styles();
-
-        fn roll_palette(rng: &mut RNG, base: &Palette, data: &LoadedData, woods: &[PaletteId], stones: &[PaletteId], roofs: &[PaletteId]) -> Palette {
-            let w = &woods[rng.rand_i32_range(0, woods.len() as i32) as usize];
-            let s = &stones[rng.rand_i32_range(0, stones.len() as i32) as usize];
-            let r = &roofs[rng.rand_i32_range(0, roofs.len() as i32) as usize];
-            base.clone()
-                .merged_with(data.palettes.get(w).expect("wood palette not found"))
-                .merged_with(data.palettes.get(s).expect("stone palette not found"))
-                .merged_with(data.palettes.get(r).expect("roof palette not found"))
-        }
-        // Densest tier first; size pool per tier (houses on the main roads,
-        // cottages on the back lanes).
-        // House + Hall on every tier for now (district wealth will refine this
-        // later). Slots that can't fit a Hall's wider frontage just skip.
-        let tiers: [(&HashSet<Point2D>, &[SizeClass]); 3] = [
-            (&arterial_band, &[SizeClass::House, SizeClass::Hall]),
-            (&collector_band, &[SizeClass::House, SizeClass::Hall]),
-            (&alley_band, &[SizeClass::House, SizeClass::Hall]),
-        ];
-
-        let mut total_buildings = 0usize;
-        let mut tier_cells = [0usize; 3];   // frontage cells found per tier
-        let mut tier_placed = [0usize; 3];  // houses placed per tier
-        let mut tier_fail = [0usize; 3];    // build_house failures per tier
-        let mut tier_short = [0usize; 3];   // chains dropped: shorter than min_front
-        let mut tier_unfit = [0usize; 3];   // slots skipped: rect didn't fit the lot
-        // DEBUG: every cell detected as frontage, per tier, so we can float a
-        // marker above it and see what the placement loop actually "sees".
-        let mut tier_frontage: [HashSet<Point2D>; 3] = Default::default();
-        // Verge cells per main-road tier (arterial, collector): the gap between
-        // the road and each house front, which we pave into a forecourt so the
-        // unavoidable set-back on a diagonal reads as a shoulder, not bare grass.
-        let mut tier_verge: [HashSet<Point2D>; 2] = Default::default();
-        for lot in &sub_blocks {
-            if lot.is_empty() { continue; }
-            let Some(mut plot) = plot_from_block(lot) else { continue; };
-
-            for (ti, (band, pool)) in tiers.iter().enumerate() {
-                let min_front = pool.iter().map(|s| *s.front_width_range().start()).min().unwrap_or(0);
-                for frontage in frontage_from_roads(lot, band) {
-                    tier_cells[ti] += frontage.cells.len();
-                    tier_frontage[ti].extend(&frontage.cells);
-                    let chain_len = frontage.cells.len() as i32;
-                    if chain_len < min_front { tier_short[ti] += 1; continue; }
-                    let mut cursor: i32 = if min_front > 1 { rng.rand_i32_range(0, min_front) } else { 0 };
-                    // Shallowest depth we'll accept on a slice that can't take the
-                    // rolled depth — lets diagonal frontage (where an axis-aligned
-                    // rect overruns the staircased ribbon) still seat a house.
-                    const MIN_FIT_DEPTH: i32 = 5;
-                    while cursor + min_front <= chain_len {
-                        let size_class = *rng.choose(pool);
-                        let fw = rng.rand_i32_range(*size_class.front_width_range().start(), *size_class.front_width_range().end() + 1);
-                        let max_depth = rng.rand_i32_range(*size_class.depth_range().start(), *size_class.depth_range().end() + 1);
-                        if cursor + fw > chain_len { cursor += 1; continue; }
-                        let chain_slice = &frontage.cells[cursor as usize..(cursor + fw) as usize];
-                        // Square-frontage bias: with the culture's square chance,
-                        // make the house a square (depth = front width) if it fits,
-                        // so it gets a dome. Guarded so a 0 bias never draws RNG.
-                        // Otherwise pick the deepest depth (down to MIN_FIT_DEPTH)
-                        // that fits, shrinking the house to hug a diagonal ribbon.
-                        let want_square = culture.square_bias() > 0
-                            && rng.percent(culture.square_bias())
-                            && plot.is_rect_usable(&rect_from_frontage(chain_slice, frontage.outward, fw));
-                        let depth = if want_square {
-                            fw
-                        } else if let Some(d) = (MIN_FIT_DEPTH..=max_depth).rev()
-                            .find(|&d| plot.is_rect_usable(&rect_from_frontage(chain_slice, frontage.outward, d)))
-                        {
-                            d
-                        } else {
-                            tier_unfit[ti] += 1; cursor += 1; continue;
-                        };
-                        let rect = rect_from_frontage(chain_slice, frontage.outward, depth);
-
-                        // Desert keeps a uniform sandstone palette; other cultures
-                        // roll wood/stone/roof variants per house for variety.
-                        let palette = match culture {
-                            Culture::Desert => base_palette.clone(),
-                            _ => roll_palette(&mut rng, &base_palette, &data, &wood_ids, &stone_ids, &roof_ids),
-                        };
-                        let roof_style = roof_styles[rng.rand_i32_range(0, roof_styles.len() as i32) as usize];
-                        let plot_bounds = synthetic_plot_bounds(&rect, frontage.outward);
-                        let footprint = Footprint::from_rect(rect);
-                        // Align the main door with the road it faces: pin the floor
-                        // (= door sill) to the height of the *nearest* road cell to
-                        // this frontage. Probe outward from every frontage cell and
-                        // keep the closest road-height hit.
-                        let road_dir = P2::from(frontage.outward);
-                        let base_lvl = {
-                            let mut best: Option<(i32, i32, P2)> = None; // (dist, height, road cell)
-                            for &c in chain_slice {
-                                for step in 1..=RIBBON_DEPTH {
-                                    let probe = c + P2::new(road_dir.x * step, road_dir.y * step);
-                                    if let Some(&y) = road_h.get(&probe) {
-                                        if best.map_or(true, |(bd, _, _)| step < bd) { best = Some((step, y, probe)); }
-                                        break;
-                                    }
-                                }
-                            }
-                            best.map(|(_, y, cell)| {
-                                // If the fronting road cell carries a half-step slab,
-                                // raise the floor one block above the slab so the door
-                                // steps down onto it instead of opening onto a lip.
-                                match slab_y_by_cell.get(&cell) {
-                                    Some(&slab_y) => slab_y + 1,
-                                    None => y,
-                                }
-                            })
-                        };
-                        let mut bctx = BuildingContext::new(culture, size_class, roof_style);
-                        bctx.base_y_override = base_lvl;
-                        let mut bctx_editor = BuildCtx::new(&mut editor, &data, &palette, &mut rng);
-                        match build_house(&mut bctx_editor, footprint, &bctx, plot_bounds).await {
-                            Ok(_) => {
-                                plot.mark_rect_used(&rect, SIDE_BUFFER_CELLS);
-                                total_buildings += 1;
-                                tier_placed[ti] += 1;
-                                // Record the verge: from each frontage cell, walk
-                                // into the block (−outward) until we reach the
-                                // house. On a straight slice this is just the
-                                // frontage row; on a diagonal it's the triangular
-                                // set-back we want to pave over.
-                                if ti < 2 {
-                                    let road_dir = P2::from(frontage.outward);
-                                    let into = P2::new(-road_dir.x, -road_dir.y);
-                                    for &c in chain_slice {
-                                        let mut p = c;
-                                        let mut guard = 0;
-                                        while !rect.contains(p) && guard < 32 {
-                                            tier_verge[ti].insert(p);
-                                            p = p + into;
-                                            guard += 1;
-                                        }
-                                    }
-                                }
-                                cursor += fw + SIDE_BUFFER_CELLS;
-                            }
-                            Err(msg) => {
-                                tier_fail[ti] += 1;
-                                log::warn!("placement build_house failed: {}", msg);
-                                cursor += 1;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        println!("Placed {} buildings across {} lots", total_buildings, sub_blocks.len());
-        println!(
-            "Per-tier [frontage cells / placed / failed] — arterial: {}/{}/{}  collector: {}/{}/{}  subdivider: {}/{}/{}",
-            tier_cells[0], tier_placed[0], tier_fail[0],
-            tier_cells[1], tier_placed[1], tier_fail[1],
-            tier_cells[2], tier_placed[2], tier_fail[2],
-        );
-        println!(
-            "Per-tier skips [short-chain / rect-unfit] — arterial: {}/{}  collector: {}/{}  subdivider: {}/{}",
-            tier_short[0], tier_unfit[0],
-            tier_short[1], tier_unfit[1],
-            tier_short[2], tier_unfit[2],
-        );
-
-        // Pave the verge: a forecourt of the road's own material in the gap
-        // between each main road and its houses, so the diagonal set-back reads
-        // as a paved shoulder. Painted at the live ground top (h-1), matching the
-        // post-flatten/foundation surface. Arterial verge = stone bricks (its
-        // road material), collector verge = cobblestone.
-        let verge_blocks = [
-            Block { id: "stone_bricks".into(), data: None, state: None },
-            Block { id: "cobblestone".into(), data: None, state: None },
-        ];
-        let mut verge_total = 0usize;
-        for (ti, cells) in tier_verge.iter().enumerate() {
-            for c in cells {
-                let h = editor.world().get_ocean_floor_height_at(*c);
-                editor.place_block(&verge_blocks[ti], Point3D::new(c.x, h - 1, c.y)).await;
-                verge_total += 1;
-            }
-        }
-        println!("Paved {} verge cells (arterial {} + collector {})", verge_total, tier_verge[0].len(), tier_verge[1].len());
-
-        // Street lighting: run last, after houses have claimed their cells, so
-        // lamps line every road's verge without landing on a building. The city
-        // generator picks the lantern type city-wide.
-        let city_rect = editor.world().world_rect_2d();
-        let city_centre = (city_rect.origin + city_rect.max()) / 2;
-        let cold = {
-            let n = editor.world().get_surface_biome_at(city_centre);
-            let n = n.name();
-            n.contains("snowy") || n.contains("frozen") || n.contains("taiga")
-        };
-        let street_lantern: crate::minecraft::Block = if cold {
-            "minecraft:soul_lantern".into()
-        } else {
-            "minecraft:lantern".into()
-        };
-        let lamps = crate::generator::paths::place_street_lights(&editor, &all_paths, &street_lantern).await;
-        println!("Placed {} street lamps", lamps.len());
-
-        editor.flush_buffer().await;
     }
 }
